requirements.txt
altair==6.0.0
appdirs==1.4.4
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
bs4==0.0.2
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
cssselect==1.4.0
fake-useragent==2.2.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
lxml_html_clean==0.4.4
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
narwhals==2.18.1
numpy==2.4.3
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.3
parse==1.21.1
pillow==12.1.1
protobuf==6.33.6
pyarrow==23.0.1
pydeck==0.9.1
pyee==11.1.1
PyMySQL==1.1.2
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
selenium==4.41.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3>=1.26.0
w3lib==2.4.0
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==10.4
wsproto==1.3.2
zipp==3.23.0
